requirements.txt
allure-pytest==2.9.42
allure-python-commons==2.9.42
appdirs==1.4.4
apturl==0.5.2
atomicwrites==1.4.0
attrs==21.2.0
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
certifi==2020.12.5
chardet==4.0.0
Click==7.0
colorama==0.4.3
command-not-found==0.3
configparser==3.5.3
constants==0.6.0
cryptography==2.8
cupshelpers==1.0
Cython==0.29.23
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.1
distro==1.4.0
distro-info===0.23ubuntu1
dnspython==1.16.0
docker==5.0.0
duplicity==0.8.12.0
entrypoints==0.3
enum34==1.1.10
fasteners==0.14.1
filelock==3.0.12
future==0.18.2
gpg===1.13.1-unknown
httplib2==0.14.0
idna==2.10
importlib-metadata==1.5.0
iniconfig==1.1.1
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
lxml==4.3.5
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.0
monotonic==1.5
more-itertools==4.2.0
namedlist==1.8
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
packaging==20.3
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
pluggy==0.13.1
protobuf==3.6.1
py==1.8.2
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.3.1
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.6
pyRFC3339==1.1
pytest==6.2.4
python-apt==2.0.0+ubuntu0.20.4.4
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytz==2019.3
pyvbox==2.1.1
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.25.1
requests-unixsocket==0.2.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.16.0
ssh-import-id==5.10
systemd-python==234
tests==0.7
toml==0.10.2
ubuntu-advantage-tools==27.0
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
vboxapi==1.0
virtualbox==2.1.1
virtualenv==20.4.7
wadllib==1.3.3
websocket-client==1.0.1
xkit==0.0.0
zipp==1.0.0
